requirements.txt
fastapi
uvicorn
mlx
mlx-lm
pydantic
python-multipart
sse-starlette
httpx
pytest
pyyaml
